Bugfixes
Some subprocesses dont run over bash any more, but instead over the
subprocess shell.
General cleaning up of code (not yet ready).
New function for checking, if the sdcard was created. If mksdcard is not
finished when the new avd is starting -which depends of the new sdcards
size - no sd will be found.